backend/requirements.txt
fastapi==0.115.2
uvicorn==0.28.0
python-dotenv==1.1.0
agno==1.5.3
websockets==14.0
pydantic==2.6.4
PyYAML==6.0.2
